[feature #11] Fix: 피드백 반영 - 숫자를 magic number가 아닌 상수값으로 관리 - 비즈니스 변경 가능성에 용이하도록 구조 변경
